requirements.txt
django
south
psycopg2
django-tastypie
